Refactor company registration logic into CompaniesService and update related controllers
