+++ Controllers/UsersController.cs
@@ -63,7 +63,7 @@ namespace adeeb.Controllers
             return Ok();
         }
         // GET: api/users/{id}
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<ActionResult<User>> GetUser(int id)
         {
             var user = await _context.Users.FindAsync(id);
