+++ Models/Company.cs
@@ -9,7 +9,7 @@ public class Company
 
     public int Id { get; set; }
     public string Name { get; set; }
-    public string LogoUrl { get; set; }
+    public string? LogoUrl { get; set; }
     public int TotalNumberOfEmployees { get; set; }
 
     [JsonConverter(typeof(JsonStringEnumConverter))]
